training_examples/requirements.txt
python-fasthtml
uvicorn>=0.29
claudette
